FEM-1728: move static method 'parse' to config class
